tools/carpview/pyproject.toml
[tool.poetry]
name = "carpview"
version = "0.1.0"
description = ""
authors = ["rincewound <aaargh@minddebugger.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyqt5 = "^5.15.9"
pyserial = "^3.5"
asyncqt = "^0.8.0"
pyzmq = "^25.1.1"
tornado = "^6.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
carpview = "carpview.main:run"